app/vectordb/vector_store.py
import faiss
import numpy as np


class VectorStore:
    """
    Simple FAISS vector database for similarity search.
    """

    def __init__(self, dimension):

        # inner product similarity
        self.index = faiss.IndexFlatIP(dimension)

    def add_embeddings(self, embeddings):

        embeddings = np.array(embeddings).astype("float32")

        self.index.add(embeddings)

    def search(self, query_embedding, k=5):

        query_embedding = np.array(query_embedding).astype("float32")

        scores, indices = self.index.search(query_embedding, k)

        return scores, indices