requirements.txt
fastapi
uvicorn
scikit-learn
sentence-transformers
faiss-cpu
numpy
pandas
scikit-fuzzy
tqdm
umap-learn